+++ web/src/recent_view_data.ts
@@ -1,6 +1,6 @@
+import type {Message} from "./message_store";
 import * as people from "./people";
 import {get_key_from_message} from "./recent_view_util";
-import type {Message} from "./types";
 
 export type ConversationData = {
     last_msg_id: number;
